pyproject.toml
[project]
name = "flaron"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.5",
    "fastapi>=0.136.0",
    "python-dotenv==1.0.0",
    "rich>=15.0.0",
    "slack-bolt>=1.28.0",
    "uvicorn>=0.44.0",
]
